requirements.txt
Flask==2.3.2
flask-cors==3.0.10
requests==2.32.5
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.3.2
xgboost==1.7.6
joblib==1.3.2
python-dotenv==1.0.0
